# Project_1/micro-loan-pwa/openapi.yaml
openapi: 3.0.3
info:
  title: Micro-Loan Calculator API
  description: |
    Backend API for the Micro-Loan Calculator PWA.
    
    Supports:
    - Loan calculations with validation
    - Cloud sync for loan history and savings goals
    - Financial literacy content delivery
    - User authentication (optional)
    - Lender database (future)
    
    PRD Reference: Sections 3.1-3.4, Phase 3 requirements
  version: 1.0.0
  contact:
    name: API Support
    email: support@microloan.app
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.microloan.app/api
    description: Production server
...
paths:
...
  /lessons:
    get:
      tags:
        - Lessons
      summary: Get all financial literacy lessons
      description: |
        Retrieve financial literacy lessons.
        PRD Reference: Section 3.3
      operationId: getLessons
      parameters:
        - name: language
          in: query
          schema:
            type: string
            default: en
            enum: [en, es, sw]
          description: Language code (en=English, es=Spanish, sw=Swahili)
      responses:
        '200':
          description: List of lessons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LessonsResponse'
  /lessons/{id}:
    get:
      tags:
        - Lessons
      summary: Get a specific lesson
      operationId: getLessonById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: language
          in: query
          schema:
            type: string
            default: en
      responses:
        '200':
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '404':
          description: Lesson not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================
  # AUTH ENDPOINTS
  # ==================
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: Optional user registration for cloud sync
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Auth
      summary: Get current user profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================
  # LENDERS ENDPOINTS (Phase 3)
  # ==================
  /lenders:
    get:
      tags:
        - Lenders
      summary: Get list of lenders
      description: |
        Lender database with reviews.
        PRD Reference: Phase 3
      operationId: getLenders
      parameters:
        - name: country
          in: query
          schema:
            type: string
          description: Filter by country code (e.g., KE, NG, IN)
        - name: minRating
          in: query
          schema:
            type: number
            minimum: 1
            maximum: 5
      responses:
        '200':
          description: List of lenders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LendersResponse'
  /lenders/{id}:
    get:
      tags:
        - Lenders
      summary: Get lender details
      operationId: getLenderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lender details with reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lender'
        '404':
          description: Lender not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================
  # SYNC ENDPOINT
  # ==================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          example: Understanding Interest Rates
        emoji:
          type: string
          example: "📈"
        summary:
          type: string
        content:
          type: string
          description: Full lesson content in markdown
        language:
          type: string
          example: en
        order:
          type: integer
          description: Display order
    LessonsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Lesson'

    # ==================
    # AUTH SCHEMAS
    # ==================
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT token
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ==================
    # LENDER SCHEMAS
    # ==================
    Lender:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Community Microfinance
        country:
          type: string
          example: KE
        averageAPR:
          type: number
          example: 24.5
        rating:
          type: number
          minimum: 1
          maximum: 5
          example: 4.2
        reviewCount:
          type: integer
          example: 156
        website:
          type: string
          format: uri
          nullable: true
        phone:
          type: string
          nullable: true
        description:
          type: string
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/Review'
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
    LendersResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Lender'

    # ==================
    # SYNC SCHEMAS
    # ==================
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: object
              additionalProperties: true
...